# docs/swagger.yaml
basePath: /
definitions:
  models.Author:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  models.Book:
    properties:
      author:
        $ref: '#/definitions/models.Author'
      author_id:
        type: integer
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: integer
      id:
        type: integer
      price:
        type: number
      title:
        type: string
    required:
    - price
    - title
    type: object
  models.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  models.CreateBookInput:
    properties:
      author_id:
        type: integer
      category_id:
        type: integer
      price:
        type: number
      title:
        type: string
    required:
    - author_id
    - category_id
    - price
    - title
    type: object
  models.UpdateBookInput:
    properties:
      author_id:
        type: integer
      category_id:
        type: integer
      price:
        type: number
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Бұл жеке медициналық клиниканың API-сы
  title: Private Medical Clinic API
  version: "1.0"
paths:
  /authors:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Author'
            type: array
      summary: Get all authors
      tags:
      - authors
    post:
      consumes:
      - application/json
      parameters:
      - description: Author
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/models.Author'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Author'
      summary: Create author
      tags:
      - authors
  /authors/{id}:
    get:
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Author'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get author by ID
      tags:
      - authors
  /books:
    get:
      consumes:
      - application/json
      description: Get books with pagination and filters
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Category ID
        in: query
        name: category
        type: integer
      - description: Author ID
        in: query
        name: author
        type: integer
      - description: Book title
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
      summary: Get all books
      tags:
      - books
    post:
      consumes:
      - application/json
      parameters:
      - description: Book
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.CreateBookInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Book'
      summary: Create book
      tags:
      - books
  /books/{id}:
    delete:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete book
      tags:
      - books
    get:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get book by ID
      tags:
      - books
    put:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated book
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBookInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
      summary: Update book
      tags:
      - books
  /categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Category'
      summary: Create category
      tags:
      - categories
swagger: "2.0"